[INJIMOB-3625] fix date format to consistent format across app
