+++ components/ui/Timestamp.tsx
@@ -1,22 +1,9 @@
 import React from 'react';
 import {View} from 'react-native';
 import {Text} from '../ui/Text';
+import { formattedDate } from '../../shared/openId4VCI/Utils';
 
 export const Timestamp: React.FC<TimestampProps> = props => {
-  function formattedDate(): React.ReactNode {
-    const date = new Date(props.time);
-    const day = date.getDate();
-    const month = date.toLocaleString('default', {month: 'long'});
-    const year = date.getFullYear();
-    const formattedHours = (date.getHours() % 12 || 12)
-      .toString()
-      .padStart(2, '0');
-    const formattedMinutes = date.getMinutes().toString().padStart(2, '0');
-    const period = date.getHours() >= 12 ? 'PM' : 'AM';
-
-    return `${day} ${month} ${year}, ${formattedHours}:${formattedMinutes} ${period}`;
-  }
-
   return (
     <View>
       <Text
@@ -29,7 +16,7 @@ export const Timestamp: React.FC<TimestampProps> = props => {
           letterSpacing: 0,
           lineHeight: 17,
         }}>
-        {formattedDate()}
+        {formattedDate(props.time)}
       </Text>
     </View>
   );
